refactoring & documentation
