+++ src/main/webapp/app/code-editor/ace/code-editor-ace.component.html
@@ -12,7 +12,7 @@
             [mode]="editorMode"
             [readOnly]="false"
             [autoUpdateContent]="true"
-            [durationBeforeCallback]="200"
+            [durationBeforeCallback]="1000"
             (textChanged)="onFileTextChanged($event)"
             class="panel panel-default"
             style="display: block; width:100%; min-height: 100px; overflow: auto; line-height: 14px;"
