+++ src/main/webapp/app/code-editor/index.ts
@@ -11,3 +11,5 @@ export * from './instructions/code-editor-instructions.component';
 export * from './instructions/code-editor-instructions-result-detail';
 export * from './status/code-editor-status.component';
 export * from './actions/code-editor-actions.component';
+export { EditorState } from 'app/entities/ace-editor/editor-state.model';
+export { CommitState } from 'app/entities/ace-editor/commit-state.model';
