+++ src/main/webapp/app/entities/ace-editor/commit-state.model.ts
@@ -1,0 +1,6 @@
+export enum CommitState {
+    CLEAN = 'CLEAN',
+    UNCOMMITTED_CHANGES = 'UNCOMMITTED_CHANGES',
+    WANTS_TO_COMMIT = 'WANTS_TO_COMMIT',
+    COMMITTING = 'COMMITTING',
+}
