+++ src/main/webapp/app/entities/ace-editor/editor-state.model.ts
@@ -1,0 +1,5 @@
+export enum EditorState {
+    CLEAN = 'CLEAN',
+    UNSAVED_CHANGES = 'UNSAVED_CHANGES',
+    SAVING = 'SAVING',
+}
